# std.crypto.onetimeauth.Ghash: make GHASH 2 - 2.5x faster