Rewrite GHASH to use 128-bit multiplication over non-reversed integers, and up to 8 blocks aggregated reduction.

`lib/std/crypto/benchmark.zig` results:

Xeon E5:

```
  Before: 1604 MiB/s
   After: 4005 MiB/s
```

Apple M1:

```
  Before: 2769 MiB/s
   After: 6014 MiB/s
```

This also makes AES-GCM faster by the way.