@eLeCtrOssSnake 2022-10-31
So basically more vectorization and better algorithms. The speed boost is insane.